Reformats URL to read better. Adds Postman Tests
The Aggregate routes are now more readable with the type being moved to
the front of the URL. Also removed some dead code from the repository
that was not in use. Added some Postman Tests as well. The collection
should be importable and includes tests to assert that the responses to
the requests are what we expect! :tada: