Companies Now Standalone object
Companies are now stored in their own table and referenced by a foreign
key `companyID` in the stocks table. Companies include their own
repository. Also updated error handling in the aggregate routes to check
if the DB is empty and return a 404 if so.

The method used for creating the companies is sort of super dirty and
quite inefficient but it works for now and uses a lot of IO to do it.
I need to get the ObjectMapper pulling out unique company symbols instead
of all 7215 of them, that would make this much more efficient.